Revert "Implement support for form owner"
This reverts commit 69f706988f7d8d79c994809f1a1f25bf9499d959.
