+++ src/rcdom.rs
@@ -223,11 +223,6 @@ impl TreeSink for RcDom {
         new_node(Comment(text))
     }
 
-    fn has_parent_node(&self, node: Handle) -> bool {
-        let node = node.borrow();
-        node.parent.is_some()
-    }
-
     fn append(&mut self, parent: Handle, child: NodeOrText<Handle>) {
         // Append to an existing Text node if we have one.
         match child {
@@ -246,9 +241,8 @@ impl TreeSink for RcDom {
 
     fn append_before_sibling(&mut self,
             sibling: Handle,
-            child: NodeOrText<Handle>) {
-        let (parent, i) = get_parent_and_index(&sibling)
-            .expect("append_before_sibling called on node without parent");
+            child: NodeOrText<Handle>) -> Result<(), NodeOrText<Handle>> {
+        let (parent, i) = unwrap_or_return!(get_parent_and_index(&sibling), Err(child));
 
         let child = match (child, i) {
             // No previous node.
@@ -259,7 +253,7 @@ impl TreeSink for RcDom {
                 let parent = parent.borrow();
                 let prev = &parent.children[i-1];
                 if append_to_existing_text(prev, &text) {
-                    return;
+                    return Ok(());
                 }
                 new_node(Text(text))
             }
@@ -277,6 +271,7 @@ impl TreeSink for RcDom {
 
         child.borrow_mut().parent = Some(Rc::downgrade(&parent));
         parent.borrow_mut().children.insert(i, child);
+        Ok(())
     }
 
     fn append_doctype_to_document(&mut self,
